LayoutTests/inspector/editor/text-editor-undo-redo.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>
function test()
{
    function textWithSelection(text, selection)
    {
        if (!selection)
            return text;

        function lineWithCursor(line, column, cursorChar)
        {
            return line.substring(0, column) + cursorChar + line.substring(column);
        }

        var lines = text.split("\n");
        selection = selection.normalize();
        var endCursorChar = selection.isEmpty() ? "|" : "<";
        lines[selection.endLine] = lineWithCursor(lines[selection.endLine], selection.endColumn, endCursorChar);
        if (!selection.isEmpty()) {
            lines[selection.startLine] = lineWithCursor(lines[selection.startLine], selection.startColumn, ">");
        }
        return lines.join("\n");
    }

    function dumpTextModel(prefix, textModel, range)
    {
        var text = textWithSelection(textModel.text(), range);
        InspectorTest.addResult(prefix + text);
    }

    function typeText(textModel, startRange, text)
    {
        var range = startRange;
        for (var i = 0; i < text.length; ++i)
            range = textModel.editRange(range, text[i]).collapseToEnd();
        return range;
    }
    InspectorTest.runTestSuite([
        function testUndoRedoTab(next)
        {
            var textModel = new WebInspector.TextEditorModel();
            textModel.setText("1\n2\n3\n");
            dumpTextModel("Text before edit:\n", textModel);
            var range = typeText(textModel, new WebInspector.TextRange(1, 0, 1, 0), "\t")
            dumpTextModel("Text after edit:\n", textModel, range);
            range = textModel.undo();
            dumpTextModel("Text after undo:\n", textModel, range);
            range = textModel.redo();
            dumpTextModel("Text after redo:\n", textModel, range);
            next();
        },

        function testConsecutiveCharactersAndNewLines(next)
        {
            var textModel = new WebInspector.TextEditorModel();
            textModel.setText("function foo()\n{\n\n}\n");
            dumpTextModel("Text before edit:\n", textModel);
            var range = typeText(textModel, new WebInspector.TextRange(2, 0, 2, 0), "    bar();\n    baz();\n    foo();");
            dumpTextModel("Text after edit:\n", textModel, range);
            range = textModel.undo();
            dumpTextModel("Text after first undo:\n", textModel, range);
            range = textModel.undo();
            dumpTextModel("Text after second undo:\n", textModel, range);
            range = textModel.undo();
            dumpTextModel("Text after third undo:\n", textModel, range);
            range = textModel.redo();
            dumpTextModel("Text after first redo:\n", textModel, range);
            range = textModel.redo();
            dumpTextModel("Text after second redo:\n", textModel, range);
            range = textModel.redo();
            dumpTextModel("Text after third redo:\n", textModel, range);
            next();
        }
    ]);
}
</script>
</head>
<body onload="runTest()">
<p>Tests undo/redo operations in the editor model.</p>
</body>
</html>
